pyproject.toml
[project]
name = "logicore"
version = "0.1.0"
description = "Enterprise AI Operating System for Logistics & Supply Chain"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "langfuse>=3.14.5",
    "langgraph>=1.0.10",
    "pydantic-settings>=2.13.1",
    "qdrant-client>=1.17.0",
    "redis>=7.2.1",
    "websockets>=16.0",
]

[tool.uv.workspace]
members = ["apps/api"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-m 'not live and not kafka'"
markers = [
    "integration: requires Docker services (qdrant, postgres, redis)",
    "e2e: full end-to-end workflow tests",
    "live: real API credentials required (Azure OpenAI + Docker), run manually only",
    "kafka: requires Kafka Docker (docker compose --profile kafka up -d), run manually",
    "redteam: security and adversarial tests",
    "evaluation: LLM quality tests (slow, needs API key)",
    "slow: tests that take >10 seconds",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.coverage.run]
source = ["apps/api/src"]

[tool.coverage.report]
show_missing = true
skip_empty = true
